handling all numpy ints in json serialization

+++ cdxlib/protocol.py
@@ -28,7 +28,7 @@ class NumpyJSONEncoder(json.JSONEncoder):
     def default(self, obj):
         if isinstance(obj, np.ndarray):
             return obj.tolist()
-        elif isinstance(obj, np.int):
+        elif isinstance(obj, np.integer):
             return int(obj)
         
 def serialize_json(obj):
